+++ app/services/worldpay_service.rb
@@ -5,8 +5,7 @@ class WorldpayService
     @url = "https://secure-test.worldpay.com/jsp/merchant/xml/paymentService.jsp"
     @username = Rails.configuration.worldpay_username
     @password = Rails.configuration.worldpay_password
-
-    assign_params(params) if params.present?
+    @params = params
   end
 
   def prepare_for_payment
@@ -27,41 +26,22 @@ class WorldpayService
   end
 
   def valid_success?
-    return false unless valid_order? && valid_params? && valid_success_payment_status?
+    worldpay_validator_service = WorldpayValidatorService.new(@order, @params)
+    return false unless worldpay_validator_service.valid_success?
 
     update_saved_data
     true
   end
 
   def valid_failure?
-    return false unless valid_order? && valid_params? && valid_failure_payment_status?
+    worldpay_validator_service = WorldpayValidatorService.new(@order, @params)
+    return false unless worldpay_validator_service.valid_failure?
 
-    @order.update_after_worldpay(@status)
+    @order.update_after_worldpay(@params[:paymentStatus])
     true
   end
 
   private
-
-  def assign_params(params)
-    @params = params
-    @order_key = params[:orderKey]
-    @amount = params[:paymentAmount]
-    @currency = params[:paymentCurrency]
-    @status = params[:paymentStatus]
-    @source = params[:source]
-    @mac = params[:mac]
-  end
-
-  def update_saved_data
-    payment = Payment.new_from_worldpay(@order)
-    payment.update_after_worldpay(@params)
-    @order.update_after_worldpay(@status)
-
-    @transient_registration.finance_details.update_balance
-    @transient_registration.finance_details.save!
-  end
-
-  # Preparing for a payment
 
   def send_request
     xml_service = WorldpayXmlService.new(@transient_registration, @order)
@@ -103,85 +83,12 @@ class WorldpayService
     Payment.new_from_worldpay(order)
   end
 
-  # Validating WorldPay responses
+  def update_saved_data
+    payment = Payment.new_from_worldpay(@order)
+    payment.update_after_worldpay(@params)
+    @order.update_after_worldpay(@params[:paymentStatus])
 
-  def valid_order?
-    return true if @order.present?
-
-    Rails.logger.error "Invalid WorldPay response: cannot find matching order for #{@order_key}"
-    false
-  end
-
-  def valid_params?
-    valid_mac? &&
-      valid_order_key_format? &&
-      valid_payment_amount? &&
-      valid_currency? &&
-      valid_source?
-  end
-
-  # For the most up-to-date guidance about validating the MAC, see:
-  # http://support.worldpay.com/support/kb/gg/corporate-gateway-guide/content/hostedintegration/securingpayments.htm
-  # We have an older implementation with some differences (for example, params aren't separated by colons).
-  def valid_mac?
-    data = [@order_key,
-            @amount,
-            @currency,
-            @status,
-            Rails.configuration.worldpay_macsecret].join
-    generated_mac = Digest::MD5.hexdigest(data).to_s
-    return true if generated_mac == @mac
-
-    Rails.logger.error "Invalid WorldPay response: MAC is invalid"
-    false
-  end
-
-  def valid_order_key_format?
-    valid_order_key = [Rails.configuration.worldpay_admin_code,
-                       Rails.configuration.worldpay_merchantcode,
-                       @order.order_code].join("^")
-    return true if @order_key == valid_order_key
-
-    Rails.logger.error "Invalid WorldPay response: order key #{@order_key} is not valid format"
-    false
-  end
-
-  def valid_payment_amount?
-    return true if @amount.to_i == @order.total_amount
-
-    Rails.logger.error "Invalid WorldPay response: #{@amount} is not valid payment amount"
-    false
-  end
-
-  def valid_currency?
-    return true if @currency == @order.currency
-
-    Rails.logger.error "Invalid WorldPay response: #{@currency} is not valid currency"
-    false
-  end
-
-  def valid_source?
-    return true if @source == "WP"
-
-    Rails.logger.error "Invalid WorldPay response: #{@source} is not valid source"
-    false
-  end
-
-  def valid_success_payment_status?
-    return true if @status == "AUTHORISED"
-
-    Rails.logger.error "Invalid WorldPay response: #{@status} is not valid payment status for success"
-    false
-  end
-
-  def valid_failure_payment_status?
-    statuses = %w[CANCELLED
-                  ERROR
-                  EXPIRED
-                  REFUSED]
-    return true if statuses.include?(@status)
-
-    Rails.logger.error "Invalid WorldPay response: #{@status} is not valid payment status for failure"
-    false
+    @transient_registration.finance_details.update_balance
+    @transient_registration.finance_details.save!
   end
 end
