+++ spec/services/worldpay_service_spec.rb
@@ -12,13 +12,9 @@ RSpec.describe WorldpayService do
   before do
     allow(Rails.configuration).to receive(:worldpay_admin_code).and_return("ADMIN_CODE")
     allow(Rails.configuration).to receive(:worldpay_merchantcode).and_return("MERCHANTCODE")
-    allow(Rails.configuration).to receive(:worldpay_macsecret).and_return("5r2zsonhn2t69s1q9jsub90l0ljrs59r")
     allow(Rails.configuration).to receive(:renewal_charge).and_return(105)
 
-    # We need to set a specific time so we know what order code to expect
-    Timecop.freeze(Time.new(2018, 1, 1)) do
-      FinanceDetails.new_finance_details(transient_registration)
-    end
+    FinanceDetails.new_finance_details(transient_registration)
   end
 
   let(:order) { transient_registration.finance_details.orders.first }
@@ -66,43 +62,48 @@ RSpec.describe WorldpayService do
   end
 
   describe "valid_success?" do
+    # Only include the params we need to test this service, as we stub the validator
     let(:params) do
       {
         orderKey: "#{Rails.configuration.worldpay_admin_code}^#{Rails.configuration.worldpay_merchantcode}^#{order.order_code}",
         paymentStatus: "AUTHORISED",
         paymentAmount: order.total_amount,
-        paymentCurrency: "GBP",
-        mac: "10c661a3e57360554675167982ca9948",
-        source: "WP",
         reg_identifier: transient_registration.reg_identifier
       }
     end
 
-    it "returns true" do
-      expect(worldpay_service.valid_success?).to eq(true)
-    end
-
     context "when the params are valid" do
       before do
-        worldpay_service.valid_success?
+        allow_any_instance_of(WorldpayValidatorService).to receive(:valid_success?).and_return(true)
+      end
+
+      it "returns true" do
+        expect(worldpay_service.valid_success?).to eq(true)
       end
 
       it "updates the payment status" do
+        worldpay_service.valid_success?
         expect(transient_registration.reload.finance_details.payments.first.world_pay_payment_status).to eq("AUTHORISED")
       end
 
       it "updates the order status" do
+        worldpay_service.valid_success?
         expect(transient_registration.reload.finance_details.orders.first.world_pay_status).to eq("AUTHORISED")
       end
 
       it "updates the balance" do
+        worldpay_service.valid_success?
         expect(transient_registration.reload.finance_details.balance).to eq(0)
       end
     end
 
     context "when the params are invalid" do
       before do
-        allow_any_instance_of(WorldpayService).to receive(:valid_params?).and_return(false)
+        allow_any_instance_of(WorldpayValidatorService).to receive(:valid_success?).and_return(false)
+      end
+
+      it "returns false" do
+        expect(worldpay_service.valid_success?).to eq(false)
       end
 
       it "does not update the order" do
@@ -116,96 +117,26 @@ RSpec.describe WorldpayService do
         expect(transient_registration.reload.finance_details.payments.count).to eq(0)
       end
     end
-
-    context "when the orderKey is in the wrong format" do
-      before do
-        params[:orderKey] = "foo#{order.order_code}"
-        # Change the MAC param to still be valid as this relies on the orderKey
-        params[:mac] = "590d7ced56f44fd472fdf563ade0730b"
-      end
-
-      it "returns false" do
-        expect(worldpay_service.valid_success?).to eq(false)
-      end
-    end
-
-    context "when the paymentStatus is invalid" do
-      before do
-        params[:paymentStatus] = "foo"
-        # Change the MAC param to still be valid as this relies on the paymentStatus
-        params[:mac] = "ecf0c84b1efa523ae847dd26cdf7b798"
-      end
-
-      it "returns false" do
-        expect(worldpay_service.valid_success?).to eq(false)
-      end
-    end
-
-    context "when the paymentAmount is invalid" do
-      before do
-        params[:paymentAmount] = 42
-        # Change the MAC param to still be valid as this relies on the paymentAmount
-        params[:mac] = "926883e7cf68b253503446d9cc50f60d"
-      end
-
-      it "returns false" do
-        expect(worldpay_service.valid_success?).to eq(false)
-      end
-    end
-
-    context "when the paymentCurrency is invalid" do
-      before do
-        params[:paymentCurrency] = "foo"
-        # Change the MAC param to still be valid as this relies on the paymentCurrency
-        params[:mac] = "838742835243dd1053e92b3b0135c905"
-      end
-
-      it "returns false" do
-        expect(worldpay_service.valid_success?).to eq(false)
-      end
-    end
-
-    context "when the mac is invalid" do
-      before do
-        params[:mac] = "foo"
-      end
-
-      it "returns false" do
-        expect(worldpay_service.valid_success?).to eq(false)
-      end
-    end
-
-    context "when the source is invalid" do
-      before do
-        params[:source] = "foo"
-      end
-
-      it "returns false" do
-        expect(worldpay_service.valid_success?).to eq(false)
-      end
-    end
   end
 
   describe "valid_failure?" do
+    # Only include the params we need to test this service, as we stub the validator
     let(:params) do
       {
         orderKey: "#{Rails.configuration.worldpay_admin_code}^#{Rails.configuration.worldpay_merchantcode}^#{order.order_code}",
         paymentStatus: "REFUSED",
         paymentAmount: order.total_amount,
-        paymentCurrency: "GBP",
-        mac: "b32f74da10bf1d9ebfd262d673e58fb9",
-        source: "WP",
         reg_identifier: transient_registration.reg_identifier
       }
     end
 
-    it "returns true" do
-      expect(worldpay_service.valid_failure?).to eq(true)
-    end
-
     context "when the params are valid" do
       before do
-        allow_any_instance_of(WorldpayService).to receive(:valid_params?).and_return(true)
+        allow_any_instance_of(WorldpayValidatorService).to receive(:valid_failure?).and_return(true)
+      end
+
+      it "returns true" do
+        expect(worldpay_service.valid_failure?).to eq(true)
       end
 
       it "updates the order status" do
@@ -216,27 +147,17 @@ RSpec.describe WorldpayService do
 
     context "when the params are invalid" do
       before do
-        allow_any_instance_of(WorldpayService).to receive(:valid_params?).and_return(false)
+        allow_any_instance_of(WorldpayValidatorService).to receive(:valid_failure?).and_return(false)
+      end
+
+      it "returns false" do
+        expect(worldpay_service.valid_failure?).to eq(false)
       end
 
       it "does not update the order" do
         unmodified_order = transient_registration.finance_details.orders.first
         worldpay_service.valid_failure?
         expect(transient_registration.reload.finance_details.orders.first).to eq(unmodified_order)
-      end
-    end
-
-    # We test most of the individual param validations when testing :valid_success?,
-    # so just test the unique ones for :valid_failure?
-    context "when the paymentStatus is invalid" do
-      before do
-        params[:paymentStatus] = "foo"
-        # Change the MAC param to still be valid as this relies on the paymentStatus
-        params[:mac] = "ecf0c84b1efa523ae847dd26cdf7b798"
-      end
-
-      it "returns false" do
-        expect(worldpay_service.valid_failure?).to eq(false)
       end
     end
   end
